+++ packages/client/package.json
@@ -17,7 +17,7 @@
     "prepare": "npm run build",
     "dev": "tsc -w",
     "lint": "eslint src/*.{ts,tsx}",
-    "test": "vitest"
+    "test": "vitest run"
   },
   "author": "Daniel Cooke <danielcooke1996@gmail.com>",
   "license": "See License in LICENSE.md",
