+++ packages/client/vitest.config.ts
@@ -7,6 +7,6 @@ export default defineConfig({
   test: {
     environment: 'jsdom',
     globals: true,
-    setupFiles: './tests/setup.js'
+    setupFiles: 'src/__tests__/setup.js',
   },
 });
